Move tests into their own modules
To avoid

- importing side-effectful modules
- executing tests multiple times
